+++ src/main/java/origami_editor/editor/App.java
@@ -13,12 +13,10 @@ import origami_editor.editor.component.BulletinBoard;
 import origami_editor.editor.databinding.*;
 import origami_editor.editor.drawing.FoldedFigure_Drawer;
 import origami_editor.editor.drawing.FoldedFigure_Worker_Drawer;
-import origami_editor.editor.export.Cp;
-import origami_editor.editor.export.Obj;
-import origami_editor.editor.export.Orh;
 import origami.folding.FoldedFigure;
 import origami_editor.editor.folded_figure.FoldedFigure_01;
 import origami_editor.editor.json.DefaultObjectMapper;
+import origami_editor.editor.service.FileSaveService;
 import origami_editor.editor.task.FoldingEstimateTask;
 import origami_editor.editor.task.TaskExecutor;
 import origami_editor.tools.KeyStrokeUtil;
@@ -28,8 +26,6 @@ import origami_editor.tools.StringOp;
 import javax.swing.*;
 import javax.swing.event.ListDataEvent;
 import javax.swing.event.ListDataListener;
-import javax.swing.filechooser.FileNameExtensionFilter;
-import javax.swing.plaf.basic.BasicFileChooserUI;
 import java.awt.*;
 import java.awt.event.*;
 import java.io.File;
@@ -60,6 +56,7 @@ public class App {
     public final CreasePattern_Worker mainCreasePatternWorker;    // Basic branch craftsman. Accepts input from the mouse.
     final Queue<Popup> popups = new ArrayDeque<>();
     private final MouseHandlerVoronoiCreate mouseHandlerVoronoiCreate = new MouseHandlerVoronoiCreate();
+    public  final FileSaveService fileSaveService;
     public LineSegmentSet lineSegmentsForFolding;//折畳み予測の最初に、ts1.Senbunsyuugou2Tensyuugou(lineSegmentsForFolding)として使う。　Ss0は、mainDrawingWorker.get_for_oritatami()かes1.get_for_select_oritatami()で得る。
     public BulletinBoard bulletinBoard = new BulletinBoard();
     // ------------------------------------------------------------------------
@@ -89,10 +86,16 @@ public class App {
         fileModel = new FileModel();
         mainCreasePatternWorker = new CreasePattern_Worker(this);
         foldedFiguresList = new DefaultComboBoxModel<>();
+
+        canvas = new Canvas(this);
+
+        fileSaveService = new FileSaveService(canvas, mainCreasePatternWorker, fileModel, applicationModel, historyStateModel, canvasModel, internalDivisionRatioModel, foldedFigureModel, gridModel, angleSystemModel, creasePatternCameraModel, foldedFiguresList);
     }
 
     public void start() {
         frame = new JFrame();
+        fileSaveService.setOwner(frame);
+
         frame.setTitle("Origami Editor " + ResourceUtil.getVersionFromManifest());//Specify the title and execute the constructor
         frame_title_0 = frame.getTitle();
         frame_title = frame_title_0;//Store title in variable
@@ -191,9 +194,7 @@ public class App {
 
         foldedFiguresList.removeAllElements();
 
-        Editor editor = new Editor(this);
-
-        canvas = editor.getCanvas();
+        Editor editor = new Editor(this, canvas);
 
         bulletinBoard.addChangeListener(e -> frame.repaint());
 
@@ -214,7 +215,7 @@ public class App {
         BottomPanel bottomPanel = editor.getBottomPanel();
         LeftPanel leftPanel = editor.getLeftPanel();
 
-        AppMenuBar appMenuBar = new AppMenuBar(this);
+        AppMenuBar appMenuBar = new AppMenuBar(this, applicationModel, fileSaveService);
 
         frame.setJMenuBar(appMenuBar);
 
@@ -235,9 +236,7 @@ public class App {
             }
         });
 
-        applicationModel.addPropertyChangeListener(e -> {
-            mainCreasePatternWorker.setData(applicationModel);
-        });
+        applicationModel.addPropertyChangeListener(e -> mainCreasePatternWorker.setData(applicationModel));
 
         applicationModel.addPropertyChangeListener(e -> {
             for (int i = 0; i < foldedFiguresList.getSize(); i++) {
@@ -318,7 +317,7 @@ public class App {
 
         fileModel.reset();
 
-        developmentView_initialization();
+        fileSaveService.developmentView_initialization();
 
         Button_shared_operation();
 
@@ -690,7 +689,7 @@ public class App {
                 case JOptionPane.YES_OPTION:
                     canvas.mouseDraggedValid = false;
                     canvas.mouseReleasedValid = false;
-                    saveFile();
+                    fileSaveService.saveFile();
 
                     TaskExecutor.stopTask();
                     System.exit(0);
@@ -704,32 +703,6 @@ public class App {
             TaskExecutor.stopTask();
             System.exit(0);
         }
-    }
-
-    // --------展開図の初期化-----------------------------
-    void developmentView_initialization() {
-        mainCreasePatternWorker.reset();
-        mainCreasePatternWorker.initialize();
-
-        //camera_of_orisen_nyuuryokuzu	の設定;
-        canvas.creasePatternCamera.setCameraPositionX(0.0);
-        canvas.creasePatternCamera.setCameraPositionY(0.0);
-        canvas.creasePatternCamera.setCameraAngle(0.0);
-        canvas.creasePatternCamera.setCameraMirror(1.0);
-        canvas.creasePatternCamera.setCameraZoomX(1.0);
-        canvas.creasePatternCamera.setCameraZoomY(1.0);
-        canvas.creasePatternCamera.setDisplayPositionX(350.0);
-        canvas.creasePatternCamera.setDisplayPositionY(350.0);
-
-        mainCreasePatternWorker.setCamera(canvas.creasePatternCamera);
-
-        canvasModel.reset();
-        internalDivisionRatioModel.reset();
-        foldedFigureModel.reset();
-
-        gridModel.reset();
-        angleSystemModel.reset();
-        creasePatternCameraModel.reset();
     }
 
     public void Button_shared_operation() {
@@ -778,374 +751,8 @@ public class App {
         }
     }
 
-    void exportFile() {
-        File exportFile = selectExportFile();
-
-        if (exportFile == null) {
-            return;
-        }
-
-        if (exportFile.getName().endsWith(".png") || exportFile.getName().endsWith(".jpg") || exportFile.getName().endsWith(".jpeg") || exportFile.getName().endsWith(".svg")) {
-            canvas.flg61 = false;
-            if ((canvasModel.getMouseMode() == MouseMode.OPERATION_FRAME_CREATE_61) && (mainCreasePatternWorker.getDrawingStage() == 4)) {
-                canvas.flg61 = true;
-                mainCreasePatternWorker.setDrawingStage(0);
-            }
-
-            fileModel.setExportImageFileName(exportFile.getAbsolutePath());
-            canvas.flg_wi = true;
-            repaintCanvas();//Necessary to not export the green border
-        } else if (exportFile.getName().endsWith(".cp")) {
-            Cp.exportFile(mainCreasePatternWorker.getSave_for_export(), exportFile);
-        } else if (exportFile.getName().endsWith(".orh")) {
-            Orh.exportFile(mainCreasePatternWorker.getSave_for_export_with_applicationModel(), exportFile);
-        }
-    }
-
-    /**
-     * Change the extension of the selected file in a fileChooser when changing the filefilter.
-     */
-    void applyFileChooserSwitchUpdate(JFileChooser fileChooser) {
-        fileChooser.addPropertyChangeListener(JFileChooser.FILE_FILTER_CHANGED_PROPERTY, e -> {
-            // Can also be AcceptAllFileFilter, then nothing should happen.
-            if (e.getNewValue() instanceof FileNameExtensionFilter) {
-                FileNameExtensionFilter filter = (FileNameExtensionFilter) e.getNewValue();
-
-                String newExtension = filter.getExtensions()[0];
-                String fileName = ((BasicFileChooserUI) fileChooser.getUI()).getFileName();
-
-                String fileBaseName = fileName;
-                if (fileName.lastIndexOf(".") > -1) {
-                    fileBaseName = fileName.substring(0, fileName.lastIndexOf("."));
-                }
-
-                fileChooser.setSelectedFile(new File(fileBaseName + "." + newExtension));
-            }
-        });
-    }
-
-    File selectOpenFile() {
-        JFileChooser fileChooser = new JFileChooser(applicationModel.getDefaultDirectory());
-        fileChooser.setDialogTitle("Open");
-
-        fileChooser.setFileFilter(new FileNameExtensionFilter("All supported files (*.ori, *.cp)", "cp", "ori"));
-        fileChooser.addChoosableFileFilter(new FileNameExtensionFilter("Origami Editor (*.ori)", "ori"));
-        fileChooser.addChoosableFileFilter(new FileNameExtensionFilter("CP / ORIPA (*.cp)", "cp"));
-
-        fileChooser.showOpenDialog(frame);
-
-        File selectedFile = fileChooser.getSelectedFile();
-        if (selectedFile != null && selectedFile.exists()) {
-            applicationModel.setDefaultDirectory(selectedFile.getParent());
-            fileModel.setSavedFileName(selectedFile.getAbsolutePath());
-            fileModel.setSaved(true);
-            historyStateModel.reset();
-
-            applicationModel.addRecentFile(selectedFile);
-
-            return selectedFile;
-        }
-
-        return null;
-    }
-
-    File selectSaveFile() {
-        JFileChooser fileChooser = new JFileChooser(applicationModel.getDefaultDirectory());
-        fileChooser.setDialogTitle("Save As");
-
-        FileNameExtensionFilter oriFilter = new FileNameExtensionFilter("Origami Editor (*.ori)", "ori");
-        fileChooser.setFileFilter(oriFilter);
-        FileNameExtensionFilter cpFilter = new FileNameExtensionFilter("CP / ORIPA (*.cp)", "cp");
-        fileChooser.addChoosableFileFilter(cpFilter);
-        fileChooser.setSelectedFile(new File("untitled.ori"));
-
-        applyFileChooserSwitchUpdate(fileChooser);
-
-        File selectedFile;
-        int choice = JOptionPane.NO_OPTION;
-        do {
-            int saveChoice = fileChooser.showSaveDialog(frame);
-
-            if (saveChoice != JFileChooser.APPROVE_OPTION) {
-                return null;
-            }
-
-            selectedFile = fileChooser.getSelectedFile();
-
-            if (selectedFile != null && !selectedFile.getName().endsWith(".ori") && !selectedFile.getName().endsWith(".cp")) {
-                if (fileChooser.getFileFilter() == oriFilter) {
-                    selectedFile = new File(selectedFile.getPath() + ".ori");
-                } else if (fileChooser.getFileFilter() == cpFilter) {
-                    selectedFile = new File(selectedFile.getPath() + ".cp");
-                }
-            }
-
-            if (selectedFile != null && selectedFile.exists()) {
-                choice = JOptionPane.showConfirmDialog(frame, "<html>File already exists.<br/>Do you want to replace it?", "Confirm Save As", JOptionPane.YES_NO_OPTION, JOptionPane.WARNING_MESSAGE);
-            }
-        } while (selectedFile != null && selectedFile.exists() && choice != JOptionPane.YES_OPTION);
-
-        if (selectedFile != null) {
-            applicationModel.setDefaultDirectory(selectedFile.getParent());
-            fileModel.setSavedFileName(selectedFile.getAbsolutePath());
-            fileModel.setSaved(true);
-            applicationModel.addRecentFile(selectedFile);
-        }
-
-        return selectedFile;
-    }
-
-    File selectImportFile() {
-        JFileChooser fileChooser = new JFileChooser(applicationModel.getDefaultDirectory());
-        fileChooser.setDialogTitle("Import");
-
-        fileChooser.setFileFilter(new FileNameExtensionFilter("All supported files", "cp", "orh", "ori"));
-        fileChooser.addChoosableFileFilter(new FileNameExtensionFilter("CP / ORIPA (*.cp)", "cp"));
-        fileChooser.addChoosableFileFilter(new FileNameExtensionFilter("Orihime (*.orh)", "orh"));
-        fileChooser.addChoosableFileFilter(new FileNameExtensionFilter("Origami Editor (*.ori)", "ori"));
-
-        fileChooser.showOpenDialog(frame);
-
-        File selectedFile = fileChooser.getSelectedFile();
-        if (selectedFile != null) {
-            applicationModel.setDefaultDirectory(selectedFile.getParent());
-        }
-
-        return selectedFile;
-    }
-
-    public File selectExportFile() {
-        JFileChooser fileChooser = new JFileChooser(applicationModel.getDefaultDirectory());
-        fileChooser.setDialogTitle("Export");
-
-        fileChooser.addChoosableFileFilter(new FileNameExtensionFilter("Image (*.png)", "png"));
-        fileChooser.addChoosableFileFilter(new FileNameExtensionFilter("Image (*.jpg)", "jpg", "jpeg"));
-        fileChooser.addChoosableFileFilter(new FileNameExtensionFilter("Image (*.svg)", "svg"));
-        fileChooser.addChoosableFileFilter(new FileNameExtensionFilter("CP / ORIPA (*.cp)", "cp"));
-        fileChooser.addChoosableFileFilter(new FileNameExtensionFilter("Orihime (*.orh)", "orh"));
-        fileChooser.setSelectedFile(new File("creasepattern.png"));
-
-        applyFileChooserSwitchUpdate(fileChooser);
-
-        File selectedFile;
-        int choice = JOptionPane.NO_OPTION;
-        do {
-            int saveChoice = fileChooser.showSaveDialog(frame);
-
-            if (saveChoice != JFileChooser.APPROVE_OPTION) {
-                return null;
-            }
-
-            selectedFile = fileChooser.getSelectedFile();
-            if (selectedFile != null && selectedFile.exists()) {
-                choice = JOptionPane.showConfirmDialog(frame, "<html>File already exists.<br/>Do you want to replace it?", "Confirm Save As", JOptionPane.YES_NO_OPTION, JOptionPane.WARNING_MESSAGE);
-            }
-        } while (selectedFile != null && selectedFile.exists() && choice == JOptionPane.NO_OPTION);
-
-        if (selectedFile != null) {
-            applicationModel.setDefaultDirectory(selectedFile.getParent());
-        }
-
-        return selectedFile;
-    }
-
-    Save readImportFile(File file) {
-        if (file == null) {
-            return null;
-        }
-
-        if (!file.exists()) {
-            return null;
-        }
-
-        Save save = null;
-
-        try {
-            if (file.getName().endsWith(".ori")) {
-                try {
-                    ObjectMapper mapper = new DefaultObjectMapper();
-                    return mapper.readValue(file, Save.class);
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-            }
-
-            if (file.getName().endsWith(".obj")) {
-                save = Obj.importFile(file);
-            }
-
-            if (file.getName().endsWith(".cp")) {
-                save = Cp.importFile(file);
-            }
-
-            if (file.getName().endsWith(".orh")) {
-                save = Orh.importFile(file);
-            }
-
-        } catch (IOException e) {
-            System.out.println(e);
-
-            JOptionPane.showMessageDialog(frame, "Opening of the saved file failed", "Opening failed", JOptionPane.ERROR_MESSAGE);
-
-            fileModel.setSavedFileName(null);
-
-            return new Save();
-        }
-
-        return save;
-    }
-
-    void saveFile() {
-        if (fileModel.getSavedFileName() == null) {
-            saveAsFile();
-
-            return;
-        }
-
-        File file = new File(fileModel.getSavedFileName());
-
-        Save save = mainCreasePatternWorker.getSave_for_export();
-        save.setVersion(ResourceUtil.getVersionFromManifest());
-
-        saveAndName2File(save, file);
-
-        fileModel.setSaved(true);
-    }
-
-    void saveAsFile() {
-        File file = selectSaveFile();
-
-        if (file == null) {
-            return;
-        }
-
-        Save save = mainCreasePatternWorker.getSave_for_export();
-        save.setVersion(ResourceUtil.getVersionFromManifest());
-
-        saveAndName2File(save, file);
-
-        fileModel.setSaved(true);
-    }
-
-    void saveAndName2File(Save save, File fname) {
-        if (fname.getName().endsWith(".ori")) {
-            try {
-                ObjectMapper mapper = new DefaultObjectMapper();
-
-                mapper.writeValue(fname, save);
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        } else if (fname.getName().endsWith(".cp")) {
-            if (!save.canSaveAsCp()) {
-                JOptionPane.showMessageDialog(frame, "The saved .cp file does not contain circles and yellow aux lines. Save as a .ori file to also save these lines.", "Warning", JOptionPane.WARNING_MESSAGE);
-            }
-
-            Cp.exportFile(save, fname);
-        } else {
-            JOptionPane.showMessageDialog(frame, "Unknown file type, cannot save", "Error", JOptionPane.ERROR_MESSAGE);
-        }
-    }
-
     public void folding_estimated(FoldedFigure_Drawer selectedFigure) throws InterruptedException, FoldingException {
         selectedFigure.folding_estimated(canvas.creasePatternCamera, lineSegmentsForFolding);
-    }
-
-    public double string2double(String str0, double default_if_error) {
-        String new_str0 = str0.trim();
-        if (new_str0.equals("L1")) {
-            str0 = String.valueOf(measuresModel.getMeasuredLength1());
-        }
-        if (new_str0.equals("L2")) {
-            str0 = String.valueOf(measuresModel.getMeasuredLength2());
-        }
-        if (new_str0.equals("A1")) {
-            str0 = String.valueOf(measuresModel.getMeasuredAngle1());
-        }
-        if (new_str0.equals("A2")) {
-            str0 = String.valueOf(measuresModel.getMeasuredAngle2());
-        }
-        if (new_str0.equals("A3")) {
-            str0 = String.valueOf(measuresModel.getMeasuredAngle3());
-        }
-
-        return StringOp.String2double(str0, default_if_error);
-    }
-
-    public void openFile(File file) {
-        if (file == null || !file.exists()) {
-            return;
-        }
-
-        fileModel.setSaved(true);
-        fileModel.setSavedFileName(file.getAbsolutePath());
-        applicationModel.setDefaultDirectory(file.getParent());
-
-        Save memo_temp = readImportFile(file);
-        System.out.println("readFile2Memo() 終了");
-
-        if (memo_temp != null) {
-            //Initialization of development drawing started
-            developmentView_initialization();
-            //Deployment parameter initialization
-
-            foldedFiguresList.removeAllElements();
-
-            mainCreasePatternWorker.setCamera(canvas.creasePatternCamera);//20170702この１行を入れると、解凍したjarファイルで実行し、最初にデータ読み込んだ直後はホイールでの展開図拡大縮小ができなくなる。jarのままで実行させた場合はもんだいないようだ。原因不明。
-            mainCreasePatternWorker.setSave_for_reading(memo_temp);
-            mainCreasePatternWorker.record();
-        }
-    }
-
-    public void openFile() {
-        System.out.println("readFile2Memo() 開始");
-
-        if (!fileModel.isSaved()) {
-            int choice = JOptionPane.showConfirmDialog(frame, "<html>Current file not saved.<br/>Do you want to save it?", "File not saved", JOptionPane.YES_NO_CANCEL_OPTION, JOptionPane.WARNING_MESSAGE);
-
-            if (choice == JOptionPane.YES_OPTION) {
-                saveFile();
-            } else if (choice == JOptionPane.CLOSED_OPTION || choice == JOptionPane.CANCEL_OPTION) {
-                return;
-            }
-        }
-
-        File file = selectOpenFile();
-
-        openFile(file);
-    }
-
-    public void importFile() {
-        if (!fileModel.isSaved()) {
-            int choice = JOptionPane.showConfirmDialog(frame, "<html>Current file not saved.<br/>Do you want to save it?", "File not saved", JOptionPane.YES_NO_CANCEL_OPTION, JOptionPane.WARNING_MESSAGE);
-
-            if (choice == JOptionPane.YES_OPTION) {
-                saveFile();
-            } else if (choice == JOptionPane.CLOSED_OPTION || choice == JOptionPane.CANCEL_OPTION) {
-                return;
-            }
-        }
-
-        System.out.println("readFile2Memo() 開始");
-        File importFile = selectImportFile();
-        Save memo_temp = readImportFile(importFile);
-        System.out.println("readFile2Memo() 終了");
-
-        if (memo_temp != null) {
-            fileModel.setSavedFileName(null);
-
-            //Initialization of development drawing started
-            developmentView_initialization();
-            //Deployment parameter initialization
-
-            foldedFiguresList.removeAllElements();
-
-            //Initialization of folding prediction map started
-
-            mainCreasePatternWorker.setCamera(canvas.creasePatternCamera);//20170702この１行を入れると、解凍したjarファイルで実行し、最初にデータ読み込んだ直後はホイールでの展開図拡大縮小ができなくなる。jarのままで実行させた場合はもんだいないようだ。原因不明。
-            mainCreasePatternWorker.setSave_for_reading(memo_temp);
-            mainCreasePatternWorker.record();
-        }
     }
 
     public void setTooltip(AbstractButton button, String key) {
